pyproject.toml
[project]
name = "pagila-data-neon-tech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
